# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/autonomys/wallet/harvest": {
            "get": {
                "description": "Harvests rewards from the autonomous wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autonomys_wallet"
                ],
                "summary": "Harvest wallet",
                "responses": {
                    "200": {
                        "description": "AutonomysHarvestWallet Done",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid SS58 address",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autonomys/wallet/list": {
            "get": {
                "description": "Lists all registered wallets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autonomys_wallet"
                ],
                "summary": "List wallets",
                "responses": {
                    "200": {
                        "description": "List of wallets in JSON format",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/data.Wallet"
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autonomys/wallet/register/{wallet}": {
            "get": {
                "description": "Registers a new wallet to the autonomys system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autonomys_wallet"
                ],
                "summary": "Register a wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Wallet successfully registered",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid SS58 address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autonomys/wallet/unregister/{wallet}": {
            "get": {
                "description": "Removes a wallet from the autonomys system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autonomys_wallet"
                ],
                "summary": "Unregister a wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet Address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet successfully removed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid wallet address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of Minotor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "In Minotor We Trust",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "data.Wallet": {
            "type": "object",
            "properties": {
                "@timestamp": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "amount": {
                    "description": "Tell Swagger to treat it as a string",
                    "type": "number"
                }
            }
        }
    }
}